requirements.txt
# Core dependencies
streamlit>=1.29.0
python-dotenv>=1.0.0

# PDF Processing
PyMuPDF>=1.23.3
pdf2image>=1.16.3
opencv-python>=4.7.0

# OCR
pytesseract>=0.3.10
python-docx>=0.8.11
docx==0.2.4

# Language models
langchain>=0.0.310
langchain-groq>=0.0.7
langchain-community>=0.0.10
transformers>=4.36.0
torch==2.3.0
torchvision==0.18.0

# Embeddings & Vector storage
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
tiktoken>=0.5.1

# Utility
numpy>=1.24.0
tqdm>=4.66.1
dataclasses>=0.6
typing-extensions>=4.5.0
